+++ lib/handler.js
